// frontend/src/app/components/model/rooms.model.ts
export interface Rooms {
  roomsId: number;
  non_ac_ward: string;
  ac_ward: string;
  ac_cabin: string;
  schedule: string;
  ccu_room: string;
  icu_room: string;
  non_ac_cabin: string;
  contact_name: string;
  contact_phone: string;
  house_no: string;
  road_no: string;
  block_no: string;
  sector_no: string;
  area: string;
  additional_notes: string;

}